# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateTalkRequestBody:
    properties:
      custom_speech_type:
        example: Özel Şirket İçi Sunum
        type: string
      duration:
        example: 5
        type: integer
      generated_text:
        type: string
      instruction:
        description: Required for update/partial_update
        example: make the staff member more helpful
        type: string
      language:
        description: For "translate" mode, this is the target language
        example: German
        type: string
      mode:
        description: '"new", "update", "partial_update", "manual_update", or "translate"'
        example: new
        type: string
      parent_id:
        description: Required for update/partial_update
        example: 1
        type: integer
      place:
        example: Airport Check-in
        type: string
      room_id:
        description: 'Optional: attach a "new" talk to a shared room (requires writer
          membership)'
        example: 1
        type: integer
      selected_text:
        description: Required only for partial_update mode
        example: Ladies and gentlemen...
        type: string
      speech_type:
        example: politician
        type: string
      topic:
        example: Checking in baggage and asking for a window seat
        type: string
    required:
    - mode
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request parameters
        type: string
    type: object
...
  handler.TalkRequestResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/handler.TalkRequestResponse'
        type: array
      created_at:
        type: string
      custom_speech_type:
        type: string
      duration:
        type: integer
      error_message:
        type: string
      generated_text:
        type: string
      id:
        type: integer
      instruction:
        type: string
      language:
        type: string
      mode:
        type: string
      parent_id:
        type: integer
      place:
        type: string
      room_id:
        type: integer
      selected_text:
        type: string
      speech_type:
        type: string
      status:
        type: string
      topic:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      version_label:
        type: string
      version_number:
        type: integer
    type: object
...
  model.TalkRequest:
    properties:
      created_at:
        type: string
      custom_speech_type:
        type: string
      duration:
        type: integer
      error_message:
        type: string
      generated_text:
        type: string
      id:
        type: integer
      instruction:
        type: string
      language:
        type: string
      mode:
        enum:
        - new
        - update
        - partial_update
        type: string
      parent_id:
        type: integer
      place:
        type: string
      room_id:
        description: nil = personal talk; set = shared inside a Room
        type: integer
      selected_text:
        type: string
      speech_type:
        type: string
      status:
        enum:
        - pending
        - processing
        - completed
        - failed
        type: string
      topic:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      version_number:
        type: integer
    type: object
info:
  contact:
    email: support@talkforge.local
    name: TalkForge Support
  description: Production-ready authentication and role verification API for TalkForge.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TalkForge API
  version: "1.0"
paths:
...
  /api/v1/talks:
    get:
      description: Returns the dialogue generation requests for the authenticated
        user, nested in a parent-child tree structure representing branches created
        by updates.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TalkRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List dialogue requests as a tree
      tags:
      - Talks
    post:
      consumes:
      - application/json
      description: Creates a pending request to generate dialogue text via Gemini.
        Mode can be "new" or "update".
      parameters:
      - description: Create Request Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTalkRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TalkRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create dialogue generation request
      tags:
      - Talks
  /api/v1/talks/{id}:
    delete:
      description: Deletes a dialogue generation request and any nested descendant
        updates.
      parameters:
      - description: Talk Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete dialogue request
      tags:
      - Talks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"